[TT-3916] Goplugins for release-4 (#3749)
* Update plugin build script with 3.0.x fixes

This is required on master, release-3-lts and release-4.0
* Restore smoke tests
* Support ComposeV2 in smoke test
* Still on buster as 1.15 is not available on bullseye
* Build go plugin from dynamically generated directory in GOPATH

+++ images/plugin-compiler/data/build.sh
@@ -1,14 +1,17 @@
 #!/bin/bash
 
-set -xe
-
 plugin_name=$1
+plugin_id=$2
+
+PLUGIN_SOURCE_PATH=${PLUGIN_SOURCE_PATH:-/plugin-source}
+PLUGIN_BUILD_PATH="/go/src/${plugin_name%.*}$plugin_id"
 
 function usage() {
     cat <<EOF
 To build a plugin:
-      $0 <plugin_name>
+      $0 <plugin_name> <plugin_id>
 
+<plugin_id> is optional
 EOF
 }
 
@@ -17,10 +20,17 @@ if [ -z "$plugin_name" ]; then
     exit 1
 fi
 
+mkdir -p $PLUGIN_BUILD_PATH
 # Plugin's vendor folder, has precendence over the cached vendor'd dependencies from tyk
 yes | cp -r $PLUGIN_SOURCE_PATH/* $PLUGIN_BUILD_PATH || true
 
 cd $PLUGIN_BUILD_PATH
+
+if [ ! -f go.mod ]; then
+    echo 'Please use gomodules.'
+    exit 1
+fi
+
 # if plugin has go.mod
 [ -f go.mod ] && [ ! -d vendor ] && go mod vendor
 rm go.mod
@@ -30,10 +40,9 @@ rm go.mod
 # We can't just copy Tyk dependencies on top of plugin dependencies, since different package versions have different file structures
 # First we need to find which deps GW already has, remove this folders, and after copy fresh versions from GW
 # ls -d /tmp/vendor/*
-
-# github.com and rest of packages have different nesting levels, so have to handle it separately
-# ls -d /tmp/vendor/github.com/*/* | sed "s|/tmp/vendor|$PLUGIN_BUILD_PATH/vendor|g" | xargs -d '\n' rm -rf
-# ls -d /tmp/vendor/*/* | sed "s|/tmp/vendor|$PLUGIN_BUILD_PATH/vendor|g" | grep -v github | xargs -d '\n' rm -rf
+## Handle if plugin has own vendor folder, and ignore error if not
+#
+#[ -d ./vendor ] && echo 'Found vendor directory, ignoring'
 
 # Copy GW dependencies
 # yes | cp -rf /tmp/vendor/* $PLUGIN_BUILD_PATH/vendor
